Moved Protobuf messages to `InteropProxy.Message`
Also put `send_message/3` in a new module
`InteropProxyWeb.ControllerHelpers`.
