+++ interop-proxy/lib/interop_proxy_web/plugs/decode_protobuf.ex
@@ -1,7 +1,7 @@
 defmodule InteropProxyWeb.Plugs.DecodeProtobuf do
   import Plug.Conn
 
-  import InteropProxyWeb.Message
+  import InteropProxy.Message
 
   def init(opts), do: opts
