+++ interop-proxy/test/interop_proxy/message_test.exs
@@ -1,0 +1,5 @@
+defmodule InteropProxy.MessageTest do
+  use ExUnit.Case, async: true
+
+  doctest InteropProxy.Message
+end
