--- interop-proxy/test/interop_proxy_web/message_test.exs
@@ -1,5 +1,0 @@
-defmodule InteropProxyWeb.MessageTest do
-  use ExUnit.Case, async: true
-
-  doctest InteropProxyWeb.Message
-end
